Make quest list generation script a Gradle task
