+++ .github/workflows/generate-quest-list.yml
@@ -10,11 +10,16 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
-      - uses: actions/setup-java@v1
+      - name: set up JDK 11
+        uses: actions/setup-java@v2
         with:
-          java-version: 11
-      - run: sudo snap install kotlin --channel=1.4/stable --classic
-      - run: kotlinc -script -Xplugin=/snap/kotlin/current/lib/kotlinx-serialization-compiler-plugin.jar .github/generate-quest-list.main.kts
+          java-version: '11'
+          distribution: 'adopt'
+
+      - name: Grant execute permission for gradlew
+        run: chmod +x gradlew
+      - name: Generate quest list
+        run: ./gradlew generateQuestList
       - uses: actions/upload-artifact@v2
         with:
           name: quest-list.csv
