+++ build.gradle.kts
@@ -50,6 +50,15 @@ tasks.register<SophoxCountValueByCountryTask>("updateChargingStationOperators") 
     minPercent = 0.1
 }
 
+tasks.register<GenerateQuestListTask>("generateQuestList") {
+    group = "streetcomplete"
+    targetFile = "$projectDir/quest-list.csv"
+    projectDirectory = projectDir
+    sourceDirectory = projectDir.resolve("app/src/main/java/de/westnordost/streetcomplete/")
+    iconsDirectory = projectDir.resolve("res/graphics/quest/")
+    noteQuestFile = sourceDirectory.resolve("data/osmnotes/notequests/OsmNoteQuestType.kt")
+}
+
 tasks.register("updateStreetCompleteData") {
     group = "streetcomplete"
     dependsOn(
@@ -57,6 +66,7 @@ tasks.register("updateStreetCompleteData") {
         // "updateChargingStationOperators",
         // "updateClothesContainerOperators",
         // "updateAtmOperators",
+        "generateQuestList",
         "app:updatePresets",
         "app:updateNsiPresets",
         "app:updateTranslations",
